Add ability to sell stock holdings
The stockholdings table needed to have the primary key reworked, since it most naturally needed to have a composite key.  As a result, a lot of the interaction with it has been modified.

The /create endpoint of holdings was also moved to buy and it now adds to existing holdings instead of creating a new one for an existing stock.